(#451) ♻️: StickerUseCaseProtocol 분리
